pyproject.toml
[project]
name = "learn-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
